Fixed linux build

+++ src/cpp/toadlet/tadpole/ResourceManager.h
@@ -46,7 +46,7 @@ public:
 	virtual void setInputStreamFactory(egg::io::InputStreamFactory *inputStreamFactory){mInputStreamFactory=inputStreamFactory;}
 
 	virtual egg::Resource::ptr get(const egg::String &name);
-	virtual egg::Resource::ptr find(const egg::String &name,ResourceHandlerData::ptr handlerData=NULL);
+	virtual egg::Resource::ptr find(const egg::String &name,ResourceHandlerData::ptr handlerData=(int)NULL);
 	virtual egg::Resource::ptr manage(const egg::Resource::ptr &resource);
 
 	virtual void unmanage(egg::Resource *resource);
